+++ adapters/I-Mobile/IMobileAdapter/GADIMobileMediatedUnifiedNativeAd.m
@@ -15,84 +15,97 @@
 #import "GADIMobileMediatedUnifiedNativeAd.h"
 #import "GADMAdapterIMobileConstants.h"
 
-/// Mapper for GADMediatedUnifiedNativeAd.
-@interface GADIMobileMediatedUnifiedNativeAd ()
+@implementation GADIMobileMediatedUnifiedNativeAd {
+  /// i-mobile native ad.
+  ImobileSdkAdsNativeObject *_iMobileNativeAd;
 
-/// i-mobile native ad.
-@property(nonatomic, strong) ImobileSdkAdsNativeObject *iMobileNativeAd;
+  /// Ad image.
+  GADNativeAdImage *_adImage;
 
-/// Ad image.
-@property(nonatomic, strong) GADNativeAdImage *adImage;
+  /// Ad image view.
+  UIImageView *_adImageView;
+}
 
-@end
-
-@implementation GADIMobileMediatedUnifiedNativeAd
-
-/// Initialize.
-- (instancetype)initWithIMobileNativeAd:(ImobileSdkAdsNativeObject *)iMobileNativeAd
-                                  image:(UIImage*)image {
-
-    // Validate arguments.
-    if (!iMobileNativeAd || !image) {
-        return nil;
-    }
-
-    // Initialize fields.
-    self = [super init];
-    if (self) {
-        self.iMobileNativeAd = iMobileNativeAd;
-        self.adImage = [[GADNativeAdImage alloc] initWithImage:image];
-    }
-    return self;
+- (nonnull instancetype)initWithIMobileNativeAd:(nonnull ImobileSdkAdsNativeObject *)iMobileNativeAd
+                                          image:(nonnull UIImage *)image {
+  // Initialize fields.
+  self = [super init];
+  if (self) {
+    _iMobileNativeAd = iMobileNativeAd;
+    _adImage = [[GADNativeAdImage alloc] initWithImage:image];
+    _adImageView = [[UIImageView alloc] initWithImage:image];
+  }
+  return self;
 }
 
 #pragma mark - GADMediatedUnifiedNativeAd
 
 - (NSString *)headline {
-    return [self.iMobileNativeAd getAdTitle];
+  return [_iMobileNativeAd getAdTitle];
 }
 
 - (NSArray<GADNativeAdImage *> *)images {
-    return @[self.adImage];
+  return @[ _adImage ];
 }
 
 - (NSString *)body {
-    return [self.iMobileNativeAd getAdDescription];
+  return [_iMobileNativeAd getAdDescription];
 }
 
 - (GADNativeAdImage *)icon {
-    return nil;
+  // Creates a 40 x 40 transparent image which acts as a placeholder image as the I-Mobile
+  // SDK does not send any image asset for the icon.
+  UIGraphicsImageRenderer *renderer =
+      [[UIGraphicsImageRenderer alloc] initWithSize:CGSizeMake(40, 40)];
+  UIImage *image =
+      [renderer imageWithActions:^(UIGraphicsImageRendererContext *_Nonnull rendererContext){
+      }];
+  return [[GADNativeAdImage alloc] initWithImage:image];
 }
 
 - (NSString *)callToAction {
-    return kGADMAdapterIMobileCallToAction;
+  return kGADMAdapterIMobileCallToAction;
 }
 
 - (NSDecimalNumber *)starRating {
-    return nil;
+  return nil;
 }
 
 - (NSString *)store {
-    return nil;
+  return nil;
 }
 
 - (NSString *)price {
-    return nil;
+  return nil;
 }
 
 - (NSString *)advertiser {
-    return [self.iMobileNativeAd getAdSponsored];
+  return [_iMobileNativeAd getAdSponsored];
 }
 
 - (NSDictionary<NSString *, id> *)extraAssets {
-    return nil;
+  return nil;
+}
+
+- (UIView *)adChoicesView {
+  return nil;
+}
+
+- (UIView *)mediaView {
+  return _adImageView;
+}
+
+- (CGFloat)mediaContentAspectRatio {
+  if (_adImageView.image.size.height > 0) {
+    return _adImageView.image.size.width / _adImageView.image.size.height;
+  }
+  return 0.0f;
 }
 
 - (void)didRecordClickOnAssetWithName:(GADUnifiedNativeAssetIdentifier)assetName
                                  view:(UIView *)view
                        viewController:(UIViewController *)viewController {
-
-    [self.iMobileNativeAd sendClick];
+  [_iMobileNativeAd sendClick];
 }
 
 @end