+++ adapters/Unity/UnityAdapter/GADMAdapterUnityConstants.h
@@ -23,7 +23,7 @@ static NSString *const kGADMAdapterUnityGameID = @"gameId";
 static NSString *const kGADMAdapterUnityPlacementID = @"zoneId";
 
 /// Ad mediation network adapter version.
-static NSString *const kGADMAdapterUnityVersion = @"3.2.0.1";
+static NSString *const kGADMAdapterUnityVersion = @"3.3.0.0";
 
 /// Ad mediation network name.
 static NSString *const kGADMAdapterUnityMediationNetworkName = @"AdMob";